[chore] #22 폴더명 변경
