+++ app/src/main/java/com/cecd/exitmed/presentation/textSearch/SearchBookmarkAdapter.kt
@@ -1,4 +1,4 @@
-package com.cecd.exitmed.presentation.search
+package com.cecd.exitmed.presentation.textSearch
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
