+++ app/src/main/java/com/cecd/exitmed/presentation/textSearch/SearchBookmarkFragment.kt
@@ -1,18 +1,17 @@
-package com.cecd.exitmed.presentation.search
+package com.cecd.exitmed.presentation.textSearch
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.viewModels
 import com.cecd.exitmed.R
-import com.cecd.exitmed.databinding.FragmentSearchHistoryBinding
+import com.cecd.exitmed.databinding.FragmentSearchBookmarkBinding
 import com.cecd.exitmed.util.binding.BindingFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class SearchHistoryFragment :
-    BindingFragment<FragmentSearchHistoryBinding>(R.layout.fragment_search_history) {
+class SearchBookmarkFragment :
+    BindingFragment<FragmentSearchBookmarkBinding>(R.layout.fragment_search_bookmark) {
     private val searchViewModel: SearchViewModel by viewModels()
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
@@ -20,8 +19,8 @@ class SearchHistoryFragment :
     }
 
     private fun initLayout() {
-        val historyAdapter = SearchHistoryAdapter()
-        binding.rvSearchHistory.adapter = historyAdapter
-        historyAdapter.setHistoryList(searchViewModel.mockHistoryList.toMutableList())
+        val adapter = SearchBookmarkAdapter()
+        binding.rvSearchBookmark.adapter = adapter
+        adapter.setBookmarkList(searchViewModel.mockBookmarkList.toMutableList())
     }
 }
