+++ app/src/main/java/com/cecd/exitmed/presentation/textSearch/SearchHistoryFragment.kt
@@ -1,17 +1,18 @@
-package com.cecd.exitmed.presentation.search
+package com.cecd.exitmed.presentation.textSearch
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.viewModels
 import com.cecd.exitmed.R
-import com.cecd.exitmed.databinding.FragmentSearchBookmarkBinding
+import com.cecd.exitmed.databinding.FragmentSearchHistoryBinding
 import com.cecd.exitmed.util.binding.BindingFragment
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class SearchBookmarkFragment :
-    BindingFragment<FragmentSearchBookmarkBinding>(R.layout.fragment_search_bookmark) {
+class SearchHistoryFragment :
+    BindingFragment<FragmentSearchHistoryBinding>(R.layout.fragment_search_history) {
     private val searchViewModel: SearchViewModel by viewModels()
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
@@ -19,8 +20,8 @@ class SearchBookmarkFragment :
     }
 
     private fun initLayout() {
-        val adapter = SearchBookmarkAdapter()
-        binding.rvSearchBookmark.adapter = adapter
-        adapter.setBookmarkList(searchViewModel.mockBookmarkList.toMutableList())
+        val historyAdapter = SearchHistoryAdapter()
+        binding.rvSearchHistory.adapter = historyAdapter
+        historyAdapter.setHistoryList(searchViewModel.mockHistoryList.toMutableList())
     }
 }
